requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-filter==23.2
django-cors-headers==4.1.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
